Refactor clause importing logic.
This commit introduces the concept of a selecting heuristic and fallback
heuristic.
The selecting heuristic selects the clauses that are better than the
given threshold.
The fallback heuristic selects its best clauses until the same number of
clauses are selected as the baseline. The baseline is determined by the
importbuffersize and importpercent.